requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
langchain==0.0.334
langchain-openai==0.0.2
openai==1.3.8
email-validator==2.1.0
pytest==7.4.3
